Backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.9
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.10.3
pytest==7.3.1
httpx==0.23.3
email-validator==1.3.1
python-dateutil==2.8.2
pydantic[email]==1.10.7
psycopg2-binary==2.9.6
